feat: use SAF browser, broaden local uploads
This brings in the changes from v0.6.0 of the Java SDK to support uploading
using InputStream objects. And with that I also changed local uploads
to use the Storage Access Framework UI instead of the Gallery. Now users
can upload basically any? local file they want to. Closes #140 and #143.

As part of the above, we're also correctly passing in MIME types now.
Closes #144. And I added the extra line needed so upload notifications
are sent at the start of an upload, not at the finish. Closes #145.